# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HinavAPI",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Default"
    }
  ],
  "paths": {
...
    "/api/libraries": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "List Libraries",
        "operationId": "list_libraries_api_libraries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GetLibrary"
                  },
                  "type": "array",
                  "title": "Response List Libraries Api Libraries Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/books": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Search Books",
        "operationId": "search_books_api_books_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uuid"
            }
          },
          {
            "name": "fileNameLike",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filenamelike"
            }
          },
          {
            "name": "cached",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cached"
            }
          },
          {
            "name": "authorLike",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorlike"
            }
          },
          {
            "name": "authorIsFavorite",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorisfavorite"
            }
          },
          {
            "name": "titleLike",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Titlelike"
            }
          },
          {
            "name": "fullText",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fulltext"
            }
          },
          {
            "name": "rate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rate"
            }
          },
          {
            "name": "seriesId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seriesid"
            }
          },
          {
            "name": "genreId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Genreid"
            }
          },
          {
            "name": "libraryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Libraryid"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sortKey",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "authors",
              "title": "Sortkey"
            }
          },
          {
            "name": "sortDesc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sortdesc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Book"
        ],
        "summary": "Update Books",
        "operationId": "update_books_api_books_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_uuid}": {
      "delete": {
        "tags": [
          "Book"
        ],
        "summary": "Delete Book",
        "operationId": "delete_book_api_books__book_uuid__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_uuid}/download": {
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "本のZipファイルダウンロード",
        "description": "指定された本のZipファイルをダウンロードする",
        "operationId": "download_book_api_books__book_uuid__download_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookAuthors": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "isFavorite": {
            "type": "boolean",
            "title": "Isfavorite",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "BookAuthors"
      },
      "BookBase": {
        "properties": {
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/BookAuthors"
            },
            "type": "array",
            "title": "Authors"
          },
          "publisher": {
            "$ref": "#/components/schemas/BookPublisher"
          },
          "isShered": {
            "type": "boolean",
            "title": "Isshered"
          },
          "chached": {
            "type": "boolean",
            "title": "Chached"
          },
          "libraryId": {
            "type": "integer",
            "title": "Libraryid"
          },
          "genreId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genreid"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/BookTag"
            },
            "type": "array",
            "title": "Tags"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "sha1": {
            "type": "string",
            "title": "Sha1"
          },
          "importFileName": {
            "type": "string",
            "title": "Importfilename"
          },
          "addDate": {
            "type": "string",
            "format": "date-time",
            "title": "Adddate"
          },
          "fileDate": {
            "type": "string",
            "format": "date-time",
            "title": "Filedate"
          },
          "userData": {
            "$ref": "#/components/schemas/BookUserDataBase"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "userId",
          "authors",
          "publisher",
          "isShered",
          "chached",
          "libraryId",
          "tags",
          "size",
          "page",
          "sha1",
          "importFileName",
          "addDate",
          "fileDate",
          "userData"
        ],
        "title": "BookBase"
      },
...
      "BookGet": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/BookBase"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "offset",
          "count",
          "rows"
        ],
        "title": "BookGet"
      },
      "BookPublisher": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "default": ""
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "title": "BookPublisher"
      },
      "BookPut": {
        "properties": {
          "uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uuids"
          },
          "seriesNo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seriesno"
          },
          "series": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "libraryId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Libraryid"
          }
        },
        "type": "object",
        "required": [
          "uuids"
        ],
        "title": "BookPut"
      },
      "BookTag": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "BookTag"
      },
      "BookUserDataBase": {
        "properties": {
          "lastOpenDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastopendate"
          },
          "readTimes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readtimes"
          },
          "openPage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openpage"
          },
          "rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          }
        },
        "type": "object",
        "title": "BookUserDataBase"
      },
...
      "GetLibrary": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          }
        },
        "type": "object",
        "required": [
          "count",
          "name",
          "id",
          "userId"
        ],
        "title": "GetLibrary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth"
          }
        }
      }
    }
  },
...
}